requirements.txt
fastapi
uvicorn
pydantic
websockets
PyNaCl>=1.5.0
